src/main/java/com/bt/codecoverage/CodeCoverageCheck.java
package com.bt.codecoverage;

public class CodeCoverageCheck {
	
	
	public boolean isPalindrome(String input) {

		if (input == null) {
			throw new IllegalArgumentException("input shouldn't be null");
		}

		if (input.equals(reverse(input))) {
			return true;
		} else {
			return false;
		}
	}

	// liril(yes) malayalam(yes)  test(x) wipe(x)
	private String reverse(String input) {
		String rev = "";
		for (int i = input.length() - 1; i >= 0; i--) {
			rev = rev + input.charAt(i);
		}
		return rev;
	}

	

}
